+++ src/Credfeto.Dispatcher.Discord/Configuration/DiscordOptions.cs
@@ -1,5 +1,8 @@
+using System.Diagnostics;
+
 namespace Credfeto.Dispatcher.Discord.Configuration;
 
+[DebuggerDisplay("WebhookUrl: {WebhookUrl}")]
 public sealed class DiscordOptions
 {
     public string WebhookUrl { get; init; } = string.Empty;